+++ apps/web/src/features/auth/model/form/index.ts
@@ -1,4 +1,5 @@
 export { useOauthAdditionalInfoForm } from './useOauthAdditionalInfoForm';
 export { useSignUpAccountStep } from './useSignUpAccountStep';
 export { useSignUpEmailVerificationStep } from './useSignUpEmailVerificationStep';
+export { useSignUpForm } from './useSignUpForm';
 export { useSignUpInfoStep } from './useSignUpInfoStep';
